Replace autogenerated stuff with some printlns to get an overview over the lifetime of a ParseController.
Also, why is metaprogramming so damn unconvenient?
`Implement every method the interface demands with a body that prints the name as well as the argument's values.` should be like 4 lines of code, instead I'm mindlessly copying every method name.
Maybe I should just have used breakpoints and stepped through it with the debugger.